BUG/MINOR: spoe: fix handling of truncated frame
The return value of decode_spoe_varint() used not to be tested before
commit a1cda0299 ("MAJOR: spoe: Add support of pipelined and asynchronous
exchanges with agents") that went into 1.8 and which introduced the test.

It's now visible in 1.7 because modern compilers emit a warning indicating
the use of a possibly uninitialized variable. But it's not just this, it's
that if any of the decode_spoe_varint() function returns -1, then returned
parse size is wrong and all the stream is desynchronized. Let's pick the
correct size check from 1.8.

There's no upstream commit ID since this was fixed by accident in 1.8.