DOC: Explicitly state only IPv4 are supported by forwardfor/originalto options
IPv6 support was added in the 2.4. However, on previous versions, it is not
explicit that only IPv4 addresses are supported. In addition, a workaround
solution is proposed to add IPv6 addresses.

This patch is related to issue #1145. There is no upstream commit ID. It may
be backported to any stable versions.

(cherry picked from commit eac279f241289b25cc469daf04e4c309e0eb0dcb)
Signed-off-by: Willy Tarreau <w@1wt.eu>