BUG/MINOR: logs: Report the true number of retries if there was no connection
When the session is aborted before any connection attempt to any server, the
number of connection retries reported in the logs is wrong. It happens
because when the retries counter is not strictly positive, we consider the
max number of retries was reached and the backend retries value is used. It
is obviously wrong when no connectioh was performed.

In fact, at this stage, the retries counter is initialized to 0. But the
backend stream-interface is in the INI state. Once it is set to SI_ST_REQ,
the counter is set to the backend value. And it is the only possible state
transition from INI state. Thus it is safe to rely on it to fix the bug.

This patch must be backported to all stable versions.

(cherry picked from commit 1d26f22e0568d6761b1e0a522079b23387bace80)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 5d5d615d891f7e45626be7348f765d31364cd358)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit de754a708ddad51839799ef9b884ae88f826689b)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit ed68cfb839cb8c20c24408c5f80325da8a4bb932)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>