BUG/MINOR: mux-h1: Release idle server H1 connection if data are received
When data are received on an idle H1 connection on server side, we must take
care to release the connection. Most of time, it will be a
408-Request-Time-out response. But, in all cases, these data should never be
processed as a response to a client.

There is no upstream ID because the 2.4 is not affected by this bug. It must
be backported as far as 2.0.

(cherry picked from commit d29ab4d43e7ec651fa034eb40228c2ec5fe101d9)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>